Movendo Enums para um arquivo único
