+++ app/model/Enum.ts
@@ -1,0 +1,86 @@
+export enum Evento {
+    EXCLUSIVIDADES = "Exclusividades",
+    JOGOS_CLASSICOS = "Jogos clássicos da EA na Epic",
+    OFERTAS_SEMANA = "Ofertas da Semana"
+}
+
+export enum Genero {
+    ACAO = "Ação",
+    ACAO_AVENTURA = "Ação e Aventura",
+    AVENTURA = "Aventura",
+    CASUAL = "Casual",
+    COMEDIA = "Comédia",
+    CONSTRUCAO_CIDADES = "Construção de Cidades",
+    CORRIDA = "Corrida",
+    CURIOSIDADES = "Curiosidades",
+    DEFESA_TORRE = "Defesa de Torre",
+    DUNGEON_CRAWLER = "Dungeon Crawler",
+    ESPACO = "Espaço",
+    ESPORTES = "Esportes",
+    ESTRATEGIA = "Estratégia",
+    ESTRATEGIA_TURNOS = "Estratégia por Turnos",
+    EXPLORACAO = "Exploração",
+    FANTASIA = "Fantasia",
+    FURTIVOS = "Furtivos",
+    INDIE = "Indie",
+    JOGO_CARTAS = "Jogo de Cartas",
+    JOGOS_FESTA = "Jogos de Festa",
+    JOGOS_TURNOS = "Jogos de Turnos",
+    LUTA = "Luta",
+    MOBA = "MOBA",
+    MUNDO_ABERTO = "Mundo Aberto",
+    MUSICA = "Música",
+    NARRACAO = "Narração",
+    PLATAFORMA = "Plataforma",
+    PRIMEIRA_PESSOA = "Primeira Pessoa",
+    QUEBRA_CABECA = "Quebra-Cabeça",
+    RETRO = "Retro",
+    ROGUELITE = "Roguelite",
+    RPG = "RPG",
+    RTS = "RTS",
+    SIMULACAO = "Simulação",
+    SOBREVIVENCIA = "Sobrevivência",
+    TERROR = "Terror",
+    TIRO = "Tiro"
+}
+
+export enum Plataforma {
+    MAC_OS = "Mac OS",
+    WINDOWS = "Windows"
+}
+
+export enum PrecoCategoria {
+    GRATUITO = "Gratuito",
+    ABAIXO_40 = "Abaixo de R$ 40,00",
+    ABAIXO_80 = "Abaixo de R$ 80,00",
+    ABAIXO_120 = "Abaixo de R$ 120,00",
+    ACIMA_59 = "R$ 59,99 e acima",
+    COM_DESCONTO = "Com desconto"
+}
+
+export enum Recurso {
+    ALEXA_GAME_CONTROL = "Alexa Game Control",
+    COMPATIVEIS_CONTROLE = "Compatíveis com Controle",
+    COMPETITIVO = "Competitivo",
+    CONQUISTAS = "Conquistas",
+    COOPERATIVO = "Cooperativo",
+    JOGOS_UM_JOGADOR = "Jogos para Um Jogador",
+    MMO = "MMO",
+    MULTIJOGADOR = "Multijogador",
+    MULTIJOGADOR_LOCAL = "Multijogador Local",
+    MULTIPLATAFORMA = "Multiplataforma",
+    ONLINE_MULTIPLAYER = "Online Multiplayer",
+    RV = "RV",
+    SALVAMENTOS_NUVEM = "Salvamentos na nuvem"
+}
+
+export enum Tipo {
+    APLICATIVOS = "Aplicativos",
+    COMPLEMENTO_JOGO = "Complemento de jogo",
+    DEMO_JOGO = "Demo do jogo",
+    EDICAO_JOGO = "Edição do Jogo",
+    EDITOR = "Editor",
+    EXPERIENCIA = "Experiência",
+    JOGO = "Jogo",
+    PACOTE_JOGO = "Pacote do Jogo"
+}
