--- app/model/Evento.ts
@@ -1,7 +1,0 @@
-export enum Evento {
-    EXCLUSIVIDADES = "Exclusividades",
-    JOGOS_CLASSICOS = "Jogos clássicos da EA na Epic",
-    OFERTAS_SEMANA = "Ofertas da Semana"
-}
-
-
