--- app/model/Genero.ts
@@ -1,41 +1,0 @@
-export enum Genero {
-    ACAO = "Ação",
-    ACAO_AVENTURA = "Ação e Aventura",
-    AVENTURA = "Aventura",
-    CASUAL = "Casual",
-    COMEDIA = "Comédia",
-    CONSTRUCAO_CIDADES = "Construção de Cidades",
-    CORRIDA = "Corrida",
-    CURIOSIDADES = "Curiosidades",
-    DEFESA_TORRE = "Defesa de Torre",
-    DUNGEON_CRAWLER = "Dungeon Crawler",
-    ESPACO = "Espaço",
-    ESPORTES = "Esportes",
-    ESTRATEGIA = "Estratégia",
-    ESTRATEGIA_TURNOS = "Estratégia por Turnos",
-    EXPLORACAO = "Exploração",
-    FANTASIA = "Fantasia",
-    FURTIVOS = "Furtivos",
-    INDIE = "Indie",
-    JOGO_CARTAS = "Jogo de Cartas",
-    JOGOS_FESTA = "Jogos de Festa",
-    JOGOS_TURNOS = "Jogos de Turnos",
-    LUTA = "Luta",
-    MOBA = "MOBA",
-    MUNDO_ABERTO = "Mundo Aberto",
-    MUSICA = "Música",
-    NARRACAO = "Narração",
-    PLATAFORMA = "Plataforma",
-    PRIMEIRA_PESSOA = "Primeira Pessoa",
-    QUEBRA_CABECA = "Quebra-Cabeça",
-    RETRO = "Retro",
-    ROGUELITE = "Roguelite",
-    RPG = "RPG",
-    RTS = "RTS",
-    SIMULACAO = "Simulação",
-    SOBREVIVENCIA = "Sobrevivência",
-    TERROR = "Terror",
-    TIRO = "Tiro"
-}
-
-
