--- app/model/Plataforma.ts
@@ -1,6 +1,0 @@
-export enum Plataforma {
-    MAC_OS = "Mac OS",
-    WINDOWS = "Windows"
-}
-
-
