--- app/model/PrecoCategoria.ts
@@ -1,10 +1,0 @@
-export enum PrecoCategoria {
-    GRATUITO = "Gratuito",
-    ABAIXO_40 = "Abaixo de R$ 40,00",
-    ABAIXO_80 = "Abaixo de R$ 80,00",
-    ABAIXO_120 = "Abaixo de R$ 120,00",
-    ACIMA_59 = "R$ 59,99 e acima",
-    COM_DESCONTO = "Com desconto"
-}
-
-
