--- app/model/Recurso.ts
@@ -1,17 +1,0 @@
-export enum Recurso {
-    ALEXA_GAME_CONTROL = "Alexa Game Control",
-    COMPATIVEIS_CONTROLE = "Compatíveis com Controle",
-    COMPETITIVO = "Competitivo",
-    CONQUISTAS = "Conquistas",
-    COOPERATIVO = "Cooperativo",
-    JOGOS_UM_JOGADOR = "Jogos para Um Jogador",
-    MMO = "MMO",
-    MULTIJOGADOR = "Multijogador",
-    MULTIJOGADOR_LOCAL = "Multijogador Local",
-    MULTIPLATAFORMA = "Multiplataforma",
-    ONLINE_MULTIPLAYER = "Online Multiplayer",
-    RV = "RV",
-    SALVAMENTOS_NUVEM = "Salvamentos na nuvem"
-}
-
-
