--- app/model/Tipo.ts
@@ -1,13 +1,0 @@
-export enum Tipo {
-    APLICATIVOS = "Aplicativos",
-    COMPLEMENTO_JOGO = "Complemento de jogo",
-    DEMO_JOGO = "Demo do jogo",
-    EDICAO_JOGO = "Edição do Jogo",
-    EDITOR = "Editor",
-    EXPERIENCIA = "Experiência",
-    JOGO = "Jogo",
-    PACOTE_JOGO = "Pacote do Jogo"
-}
-
-
-
